Moving tests into subfolder
