+++ test/cookiemask_test.go
@@ -1,6 +1,7 @@
-package pullcord
+package test
 
 import (
+	. ".."
 	"encoding/hex"
 	"errors"
 	"fmt"
