+++ test/passthru_test.go
@@ -1,6 +1,7 @@
-package pullcord
+package test
 
 import (
+	. ".."
 	"github.com/fitstar/falcore"
 	"github.com/stretchr/testify/assert"
 	"io/ioutil"
